LIU-455: Move functions from simple.py to simple_functions.py.
